backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
starlette
pydantic
python-dotenv
pypdf
pymupdf
python-docx
openpyxl
beautifulsoup4
lxml
requests
groq